feat(cv): Open cv in current windown now that it includes navbar
While the CV was previously fully separate to the to the rest of the page
it is now styled like everything else when viewed on a screen, and
features the navbar as well. Thus it no longer breaks navigation and
it's less convenient that it opens separately than in the current window.